refactor(core): split ModelCatalog types, add streaming base, DX and storage hygiene
- Extract LatencyTier/Pricing/Option/Tag into ModelCatalogTypes.swift; ModelCatalog.swift now holds only catalogue arrays
- Add BaseStreamingLiveController shared WebSocket lifecycle (enqueue/flush/stopping) for future provider dedupe
- Makefile: add lint/format/verify-checksums targets; add scripts/verify-checksums.sh for CTranscribe and baseline drift
- SecureAppStorage: document AppSettings registry coupling and add InMemoryIdentifierRegistry test double
