+++ Makefile
@@ -65,6 +65,21 @@ install-verify: release ## Build release, install to /Applications, launch and v
 .PHONY: preflight
 preflight: test-all verify ## Full pre-release check (tests + launch verification)
 
+.PHONY: lint
+lint: ## Run SwiftLint (strict) and SwiftFormat (lint mode)
+	swift package plugin --allow-writing-to-package-directory swiftlint --strict
+	swift package plugin --allow-writing-to-package-directory swiftformat --lint --recursive
+
+.PHONY: format
+format: ## Auto-fix formatting and lint violations where possible
+	swift package plugin --allow-writing-to-package-directory swiftformat --recursive
+	swift package plugin --allow-writing-to-package-directory swiftlint --fix --format
+
+.PHONY: verify-checksums
+verify-checksums: ## Verify binary XCFramework and package checksums
+	chmod +x scripts/verify-checksums.sh
+	./scripts/verify-checksums.sh
+
 .PHONY: install-hooks
 install-hooks: ## Install git hooks for pre-push verification
 	git config core.hooksPath .githooks
